Moved http configurations around.
