+++ pkg/httpconfig/config.go
@@ -1,7 +1,7 @@
 // Copyright (c) The Thanos Authors.
 // Licensed under the Apache License 2.0.
 
-package query
+package httpconfig
 
 import (
 	"fmt"
@@ -11,20 +11,20 @@ import (
 	"gopkg.in/yaml.v2"
 
 	"github.com/pkg/errors"
-	http_util "github.com/thanos-io/thanos/pkg/http"
 )
 
+// Config is a structure that allows pointing to various HTTP endpoint, e.g ruler connecting to queriers.
 type Config struct {
-	HTTPClientConfig http_util.ClientConfig    `yaml:"http_config"`
-	EndpointsConfig  http_util.EndpointsConfig `yaml:",inline"`
+	HTTPClientConfig ClientConfig    `yaml:"http_config"`
+	EndpointsConfig  EndpointsConfig `yaml:",inline"`
 }
 
 func DefaultConfig() Config {
 	return Config{
-		EndpointsConfig: http_util.EndpointsConfig{
+		EndpointsConfig: EndpointsConfig{
 			Scheme:          "http",
 			StaticAddresses: []string{},
-			FileSDConfigs:   []http_util.FileSDConfig{},
+			FileSDConfigs:   []FileSDConfig{},
 		},
 	}
 }
@@ -45,12 +45,12 @@ func LoadConfigs(confYAML []byte) ([]Config, error) {
 	return queryCfg, nil
 }
 
-// BuildQueryConfig returns a query client configuration from a static address.
-func BuildQueryConfig(queryAddrs []string) ([]Config, error) {
-	configs := make([]Config, 0, len(queryAddrs))
-	for i, addr := range queryAddrs {
+// BuildConfig returns a configuration from a static addresses.
+func BuildConfig(addrs []string) ([]Config, error) {
+	configs := make([]Config, 0, len(addrs))
+	for i, addr := range addrs {
 		if addr == "" {
-			return nil, errors.Errorf("static querier address cannot be empty at index %d", i)
+			return nil, errors.Errorf("static address cannot be empty at index %d", i)
 		}
 		// If addr is missing schema, add http.
 		if !strings.Contains(addr, "://") {
@@ -61,10 +61,10 @@ func BuildQueryConfig(queryAddrs []string) ([]Config, error) {
 			return nil, errors.Wrapf(err, "failed to parse addr %q", addr)
 		}
 		if u.Scheme != "http" && u.Scheme != "https" {
-			return nil, errors.Errorf("%q is not supported scheme for querier address", u.Scheme)
+			return nil, errors.Errorf("%q is not supported scheme for address", u.Scheme)
 		}
 		configs = append(configs, Config{
-			EndpointsConfig: http_util.EndpointsConfig{
+			EndpointsConfig: EndpointsConfig{
 				Scheme:          u.Scheme,
 				StaticAddresses: []string{u.Host},
 				PathPrefix:      u.Path,
