+++ pkg/httpconfig/config_test.go
@@ -1,16 +1,15 @@
 // Copyright (c) The Thanos Authors.
 // Licensed under the Apache License 2.0.
 
-package query
+package httpconfig
 
 import (
 	"testing"
 
-	"github.com/thanos-io/thanos/pkg/http"
 	"github.com/thanos-io/thanos/pkg/testutil"
 )
 
-func TestBuildQueryConfig(t *testing.T) {
+func TestBuildConfig(t *testing.T) {
 	for _, tc := range []struct {
 		desc      string
 		addresses []string
@@ -21,7 +20,7 @@ func TestBuildQueryConfig(t *testing.T) {
 			desc:      "single addr without path",
 			addresses: []string{"localhost:9093"},
 			expected: []Config{{
-				EndpointsConfig: http.EndpointsConfig{
+				EndpointsConfig: EndpointsConfig{
 					StaticAddresses: []string{"localhost:9093"},
 					Scheme:          "http",
 				},
@@ -32,13 +31,13 @@ func TestBuildQueryConfig(t *testing.T) {
 			addresses: []string{"localhost:9093", "localhost:9094/prefix"},
 			expected: []Config{
 				{
-					EndpointsConfig: http.EndpointsConfig{
+					EndpointsConfig: EndpointsConfig{
 						StaticAddresses: []string{"localhost:9093"},
 						Scheme:          "http",
 					},
 				},
 				{
-					EndpointsConfig: http.EndpointsConfig{
+					EndpointsConfig: EndpointsConfig{
 						StaticAddresses: []string{"localhost:9094"},
 						Scheme:          "http",
 						PathPrefix:      "/prefix",
@@ -50,7 +49,7 @@ func TestBuildQueryConfig(t *testing.T) {
 			desc:      "single addr with path and http scheme",
 			addresses: []string{"http://localhost:9093"},
 			expected: []Config{{
-				EndpointsConfig: http.EndpointsConfig{
+				EndpointsConfig: EndpointsConfig{
 					StaticAddresses: []string{"localhost:9093"},
 					Scheme:          "http",
 				},
@@ -60,7 +59,7 @@ func TestBuildQueryConfig(t *testing.T) {
 			desc:      "single addr with path and https scheme",
 			addresses: []string{"https://localhost:9093"},
 			expected: []Config{{
-				EndpointsConfig: http.EndpointsConfig{
+				EndpointsConfig: EndpointsConfig{
 					StaticAddresses: []string{"localhost:9093"},
 					Scheme:          "https",
 				},
@@ -83,7 +82,7 @@ func TestBuildQueryConfig(t *testing.T) {
 		},
 	} {
 		t.Run(tc.desc, func(t *testing.T) {
-			cfg, err := BuildQueryConfig(tc.addresses)
+			cfg, err := BuildConfig(tc.addresses)
 			if tc.err {
 				testutil.NotOk(t, err)
 				return
